moved cli files to separate folder
