+++ src/proq/cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli import main
+
+__all__ = [main]
